refactor: extract markdown helpers
